backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic-settings==2.7.1
httpx==0.28.1
PyJWT==2.10.1
beautifulsoup4==4.12.3
websockets==14.1
mplfinance==0.12.10b0
matplotlib==3.10.0
